ScadaComm: change namespace of UI classes
