+++ ScadaComm/ScadaCommCommon/Channels/UI/FrmCommSerialProps.resx
