+++ ScadaComm/ScadaCommCommon/Channels/UI/FrmCommTcpClientProps.resx
