+++ ScadaComm/ScadaCommCommon/Channels/UI/FrmCommTcpServerProps.resx
