+++ ScadaComm/ScadaCommCommon/Channels/UI/FrmCommUdpProps.cs
@@ -16,7 +16,7 @@
  * 
  * Product  : Rapid SCADA
  * Module   : ScadaCommCommon
- * Summary  : TCP client communication channel properties
+ * Summary  : UDP communication channel properties
  * 
  * Author   : Mikhail Shiryaev
  * Created  : 2015
@@ -28,22 +28,22 @@ using System;
 using System.Collections.Generic;
 using System.Windows.Forms;
 
-namespace Scada.Comm.Channels
+namespace Scada.Comm.Channels.UI
 {
     /// <summary>
-    /// TCP client communication channel properties
-    /// <para>Свойства канала связи TCP-клиент</para>
+    /// UDP communication channel properties
+    /// <para>Свойства канала связи UDP</para>
     /// </summary>
-    internal partial class FrmCommTcpClientProps : Form
+    internal partial class FrmCommUdpProps : Form
     {
         private SortedList<string, string> commCnlParams; // параметры канала связи
         private bool modified;                            // признак изменения параметров
-        private CommTcpClientLogic.Settings settings;     // настройки канала связи
+        private CommUdpLogic.Settings settings;           // настройки канала связи
 
         /// <summary>
         /// Конструктор, ограничивающий создание формы без параметров
         /// </summary>
-        private FrmCommTcpClientProps()
+        private FrmCommUdpProps()
         {
             InitializeComponent();
         }
@@ -56,7 +56,7 @@ namespace Scada.Comm.Channels
             if (commCnlParams == null)
                 throw new ArgumentNullException("commCnlParams");
 
-            FrmCommTcpClientProps form = new FrmCommTcpClientProps();
+            FrmCommUdpProps form = new FrmCommUdpProps();
             form.commCnlParams = commCnlParams;
             form.modified = false;
             form.ShowDialog();
@@ -64,21 +64,22 @@ namespace Scada.Comm.Channels
         }
 
 
-        private void FrmCommTcpClientProps_Load(object sender, EventArgs e)
+        private void FrmCommUdpProps_Load(object sender, EventArgs e)
         {
             // перевод формы
-            Translator.TranslateForm(this, "Scada.Comm.Channels.FrmCommTcpClientProps", toolTip);
+            Translator.TranslateForm(this, "Scada.Comm.Channels.FrmCommUdpProps", toolTip);
 
             // инициализация настроек канала связи
-            settings = new CommTcpClientLogic.Settings();
+            settings = new CommUdpLogic.Settings();
             settings.Init(commCnlParams, false);
 
             // установка элементов управления в соответствии с параметрами канала связи
             cbBehavior.Text = settings.Behavior.ToString();
-            cbConnMode.SetSelectedItem(settings.ConnMode, 
-                new Dictionary<string, int>() { { "Individual", 0 }, { "Shared", 1 } });
-            txtIpAddress.Text = settings.IpAddress;
-            numTcpPort.SetValue(settings.TcpPort);
+            cbDevSelMode.SetSelectedItem(settings.DevSelMode, 
+                new Dictionary<string, int>() { { "ByIPAddress", 0 }, { "ByDeviceLibrary", 1 } });
+            numLocalUdpPort.SetValue(settings.LocalUdpPort);
+            numRemoteUdpPort.SetValue(settings.RemoteUdpPort);
+            txtRemoteIpAddress.Text = settings.RemoteIpAddress;
 
             modified = false;
         }
@@ -88,9 +89,9 @@ namespace Scada.Comm.Channels
             modified = true;
         }
 
-        private void cbConnMode_SelectedIndexChanged(object sender, EventArgs e)
+        private void cbBehavior_SelectedIndexChanged(object sender, EventArgs e)
         {
-            txtIpAddress.Enabled = cbConnMode.SelectedIndex == 1; // Shared
+            cbDevSelMode.Enabled = cbBehavior.SelectedIndex == 1; // Slave
             modified = true;
         }
 
@@ -100,22 +101,18 @@ namespace Scada.Comm.Channels
             if (modified)
             {
                 settings.Behavior = cbBehavior.ParseText<CommChannelLogic.OperatingBehaviors>();
-                settings.ConnMode = (CommTcpChannelLogic.ConnectionModes)cbConnMode.GetSelectedItem(
+                settings.DevSelMode = (CommUdpLogic.DeviceSelectionModes)cbDevSelMode.GetSelectedItem(
                     new Dictionary<int, object>() { 
-                        { 0, CommTcpChannelLogic.ConnectionModes.Individual }, 
-                        { 1, CommTcpChannelLogic.ConnectionModes.Shared } });
-                settings.IpAddress = txtIpAddress.Text;
-                settings.TcpPort = Convert.ToInt32(numTcpPort.Value);
+                        { 0, CommUdpLogic.DeviceSelectionModes.ByIPAddress }, 
+                        { 1, CommUdpLogic.DeviceSelectionModes.ByDeviceLibrary } });
+                settings.LocalUdpPort = Convert.ToInt32(numLocalUdpPort.Value);
+                settings.RemoteUdpPort = Convert.ToInt32(numRemoteUdpPort.Value);
+                settings.RemoteIpAddress = txtRemoteIpAddress.Text;
 
                 settings.SetCommCnlParams(commCnlParams);
             }
 
-            // проверка настроек
-            if (settings.ConnMode == CommTcpChannelLogic.ConnectionModes.Shared &&
-                string.IsNullOrWhiteSpace(settings.IpAddress))
-                ScadaUiUtils.ShowError(CommPhrases.IpAddressRequired);
-            else
-                DialogResult = DialogResult.OK;
+            DialogResult = DialogResult.OK;
         }
 
         private void btnCancel_Click(object sender, EventArgs e)
