+++ ScadaComm/ScadaCommCommon/Channels/UI/FrmCommUdpProps.resx
